Data.Reify.Graph

Description

This is the shared definition of a Graph in Data.Reify.

Documentation

data Graph e

Graph is a basic graph structure over nodes of the higher kind e, with a single root. There is an assumption that there is no Unique used in a node which does not have a corresponding entry is the association list. The idea with this structure is that it is trivial to convert into an Array, IntMap, or into a Martin Erwig's Functional Graph, as required.

Constructors

Graph [(Unique, e Unique)] Unique

type Unique = Int